+++ core/include/core/game.hpp
@@ -4,8 +4,6 @@
 #include <span>
 
 std::filesystem::path getMainBinDir();
-
-bool* findCheatsEnabledVariable(const std::span<uint8_t> range);
 void setupCheats();
 bool isGameDllLoaded();
 void setGameWorkingDirectory();